Removed claims presentation
