+++ Projects/Claims/Sources/ClaimStore.swift
@@ -1,5 +1,5 @@
 import Apollo
-import Presentation
+import StoreContainer
 import SwiftUI
 import hCore
 import hCoreUI
