+++ Projects/Claims/Sources/Navigation/ClaimsNavigation.swift
@@ -1,5 +1,5 @@
 import Combine
-import Presentation
+import StoreContainer
 import SwiftUI
 import hCore
 import hCoreUI
@@ -148,51 +148,53 @@ public struct ClaimsNavigation: View {
         }
         .environmentObject(claimsNavigationVm)
         .onAppear {
-            let store: SubmitClaimStore = globalPresentableStoreContainer.get()
-            cancellable = store.actionSignal.publisher.sink { _ in
-            } receiveValue: { action in
-                switch action {
-                case .dismissNewClaimFlow:
-                    router.dismiss()
-                case let .navigationAction(navigationAction):
-                    switch navigationAction {
-                    case .openTriagingEntrypointScreen:
-                        router.push(ClaimsRouterActions.triagingEntrypoint)
-                    case .openTriagingOptionScreen:
-                        router.push(ClaimsRouterActions.triagingOption)
-                    case let .openDateOfOccurrencePlusLocationScreen(option):
-                        router.push(ClaimsRouterActions.dateOfOccurrancePlusLocation(option: option))
-                    case .openSelectContractScreen:
-                        router.push(ClaimsRouterActions.selectContract)
-                    case let .openPhoneNumberScreen(model):
-                        router.push(ClaimsRouterActions.phoneNumber(model: model))
-                    case .openAudioRecordingScreen:
-                        router.push(ClaimsRouterActions.audioRecording)
-                    case .openSingleItemScreen:
-                        router.push(ClaimsRouterActions.singleItem)
-                    case .openSummaryScreen:
-                        router.push(ClaimsRouterActions.summary)
-                    case let .openDeflectScreen(type):
-                        router.push(ClaimsRouterActions.deflect(type: type))
-                    case .openConfirmEmergencyScreen:
-                        router.push(ClaimsRouterActions.emergencySelect)
-                    case .openFileUploadScreen:
-                        router.push(ClaimsRouterActions.uploadFiles)
-                    case .openClaimCheckoutScreen:
-                        router.push(ClaimsRouterActions.checkOutNoRepair)
-                    case .openSuccessScreen:
-                        router.push(ClaimsRouterActionsWithoutBackButton.success)
-                    case .openFailureSceen:
-                        router.push(ClaimsRouterActionsWithoutBackButton.failure)
-                    case .openUpdateAppScreen:
-                        router.push(ClaimsRouterActionsWithoutBackButton.updateApp)
+            let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
+            cancellable = store.actionSignal
+                .receive(on: RunLoop.main)
+                .sink { _ in
+                } receiveValue: { action in
+                    switch action {
+                    case .dismissNewClaimFlow:
+                        router.dismiss()
+                    case let .navigationAction(navigationAction):
+                        switch navigationAction {
+                        case .openTriagingEntrypointScreen:
+                            router.push(ClaimsRouterActions.triagingEntrypoint)
+                        case .openTriagingOptionScreen:
+                            router.push(ClaimsRouterActions.triagingOption)
+                        case let .openDateOfOccurrencePlusLocationScreen(option):
+                            router.push(ClaimsRouterActions.dateOfOccurrancePlusLocation(option: option))
+                        case .openSelectContractScreen:
+                            router.push(ClaimsRouterActions.selectContract)
+                        case let .openPhoneNumberScreen(model):
+                            router.push(ClaimsRouterActions.phoneNumber(model: model))
+                        case .openAudioRecordingScreen:
+                            router.push(ClaimsRouterActions.audioRecording)
+                        case .openSingleItemScreen:
+                            router.push(ClaimsRouterActions.singleItem)
+                        case .openSummaryScreen:
+                            router.push(ClaimsRouterActions.summary)
+                        case let .openDeflectScreen(type):
+                            router.push(ClaimsRouterActions.deflect(type: type))
+                        case .openConfirmEmergencyScreen:
+                            router.push(ClaimsRouterActions.emergencySelect)
+                        case .openFileUploadScreen:
+                            router.push(ClaimsRouterActions.uploadFiles)
+                        case .openClaimCheckoutScreen:
+                            router.push(ClaimsRouterActions.checkOutNoRepair)
+                        case .openSuccessScreen:
+                            router.push(ClaimsRouterActionsWithoutBackButton.success)
+                        case .openFailureSceen:
+                            router.push(ClaimsRouterActionsWithoutBackButton.failure)
+                        case .openUpdateAppScreen:
+                            router.push(ClaimsRouterActionsWithoutBackButton.updateApp)
+                        default:
+                            break
+                        }
                     default:
                         break
                     }
-                default:
-                    break
                 }
-            }
         }
         .detent(
             presented: $claimsNavigationVm.isLocationPickerPresented,
@@ -248,11 +250,11 @@ public struct ClaimsNavigation: View {
     private func showClaimEntrypointGroup(origin: ClaimsOrigin) -> some View {
         SelectClaimEntrypointGroup(
             selectedEntrypoints: { entrypoints in
-                let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                 store.send(.setSelectedEntrypoints(entrypoints: entrypoints))
             })
             .onAppear {
-                let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                 store.send(.fetchEntrypointGroups)
             }
             .resetProgressToPreviousValueOnDismiss
@@ -262,7 +264,7 @@ public struct ClaimsNavigation: View {
 
     private func showClaimEntrypointType() -> some View {
         SelectClaimEntrypointType(selectedEntrypointOptions: { options, selectedEntrypointId in
-            let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+            let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
             store.send(.setSelectedEntrypointOptions(entrypoints: options, entrypointId: selectedEntrypointId))
         })
         .resetProgressToPreviousValueOnDismiss
@@ -272,7 +274,7 @@ public struct ClaimsNavigation: View {
     private func showClaimEntrypointOption() -> some View {
         SelectClaimEntrypointOption(
             onButtonClick: { entrypointId, entrypointOptionId in
-                let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                 store.send(
                     .startClaimRequest(
                         entrypointId: entrypointId,
@@ -305,7 +307,7 @@ public struct ClaimsNavigation: View {
     }
 
     private func openAudioRecordingSceen() -> some View {
-        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+        let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
         let url = store.state.audioRecordingStep?.getUrl()
         return SubmitClaimAudioRecordingScreen(url: url)
             .resetProgressToPreviousValueOnDismiss
@@ -331,7 +333,7 @@ public struct ClaimsNavigation: View {
 
     @ViewBuilder
     private func openDeflectStepScreen() -> some View {
-        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+        let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
         let model = store.state.deflectStepModel
 
         Group {
@@ -361,7 +363,7 @@ public struct ClaimsNavigation: View {
 
     private func openEmergencySelectScreen() -> some View {
         SumitClaimEmergencySelectScreen(title: {
-            let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+            let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
             return store.state.emergencyConfirm?.text ?? ""
         })
         .resetProgressToPreviousValueOnDismiss
@@ -369,7 +371,7 @@ public struct ClaimsNavigation: View {
     }
 
     private func openFileUploadScreen() -> some View {
-        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+        let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
         return SubmitClaimFilesUploadScreen(model: store.state.fileUploadStep!)
             .resetProgressToPreviousValueOnDismiss
             .addDismissClaimsFlow()
@@ -387,14 +389,14 @@ public struct ClaimsNavigation: View {
             buttons: .init(
                 actionButton: .init(
                     buttonAction: {
-                        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                        let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                         store.send(.popClaimFlow)
                     }
                 ),
                 dismissButton: .init(
                     buttonTitle: L10n.openChat,
                     buttonAction: {
-                        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                        let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                         store.send(.dismissNewClaimFlow)
                         DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {
                             store.send(.submitClaimOpenFreeTextChat)
@@ -425,7 +427,7 @@ public struct ClaimsNavigation: View {
 
     private func openPriceInputScreen() -> some View {
         PriceInputScreen(onSave: { purchasePrice in
-            let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+            let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
             store.send(.setPurchasePrice(priceOfPurchase: Double(purchasePrice)))
             claimsNavigationVm.isPriceInputPresented = false
         })
@@ -436,12 +438,12 @@ public struct ClaimsNavigation: View {
         ItemPickerScreen<ClaimFlowItemProblemOptionModel>(
             config: .init(
                 items: {
-                    let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                    let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                     return store.state.singleItemStep?.availableItemProblems
                         .compactMap({ (object: $0, displayName: .init(title: $0.displayName)) }) ?? []
                 }(),
                 preSelectedItems: {
-                    let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                    let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                     if let singleItemStep = store.state.singleItemStep {
                         let preselected = singleItemStep.availableItemProblems
                             .filter { model in
@@ -463,7 +465,7 @@ public struct ClaimsNavigation: View {
                         }
                     }
                     claimsNavigationVm.isDamagePickerPresented = false
-                    let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                    let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
                     store.send(
                         .setSingleItemDamage(
                             damages: damages
@@ -522,10 +524,10 @@ extension View {
     }
 
     var resetProgressToPreviousValueOnDismiss: some View {
-        let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+        let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
         let previousProgress = store.state.previousProgress
         return self.onDeinit {
-            let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+            let store: SubmitClaimStore = hGlobalPresentableStoreContainer.get()
             store.send(.setProgress(progress: previousProgress))
         }
     }
