+++ Projects/Claims/Sources/Views/ClaimDetailView.swift
@@ -4,7 +4,6 @@ import Home
 import Kingfisher
 import Payment
 import Photos
-import Presentation
 import StoreContainer
 import SwiftUI
 import hCore
@@ -107,7 +106,7 @@ public struct ClaimDetailView: View {
         }
         .modally(item: $vm.showFilesView) { [weak vm] item in
             ClaimFilesView(endPoint: item.endPoint, files: item.files) { _ in
-                let claimStore: ClaimsStore = globalPresentableStoreContainer.get()
+                let claimStore: ClaimsStore = hGlobalPresentableStoreContainer.get()
                 claimStore.send(.fetchClaims)
                 DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
                     let nav = UIApplication.shared.getTopViewControllerNavigation()
@@ -368,7 +367,7 @@ struct ClaimDetailView_Previews: PreviewProvider {
 }
 
 public class ClaimDetailViewModel: ObservableObject {
-    @PresentableStore var store: ClaimsStore
+    @hPresentableStore var store: ClaimsStore
     @Published var claim: ClaimModel
     var claimService = hFetchClaimService()
     @Published var fetchFilesError: String?
@@ -383,13 +382,13 @@ public class ClaimDetailViewModel: ObservableObject {
         claim: ClaimModel
     ) {
         self.claim = claim
-        let store: ClaimsStore = globalPresentableStoreContainer.get()
+        let store: ClaimsStore = hGlobalPresentableStoreContainer.get()
         let files = store.state.files[claim.id] ?? []
         self.fileGridViewModel = .init(files: files, options: [])
         Task {
             await fetchFiles()
         }
-        store.actionSignal.publisher
+        store.actionSignal
             .filter { action in
                 if case .refreshFiles = action {
                     return true
@@ -423,7 +422,7 @@ public class ClaimDetailViewModel: ObservableObject {
                 .removeDuplicates()
                 .receive(on: RunLoop.main)
                 .sink { [weak self] conversationsTimeStamp in
-                    let claimStore: ClaimsStore = globalPresentableStoreContainer.get()
+                    let claimStore: ClaimsStore = hGlobalPresentableStoreContainer.get()
                     let conversation = claimStore.state.claim(for: self?.claim.id ?? "")?.conversation
                     let conversationId = conversation?.id
                     let timeStamp = conversationsTimeStamp[conversationId ?? ""]
@@ -438,15 +437,15 @@ public class ClaimDetailViewModel: ObservableObject {
                 }
                 .store(in: &self.cancellables)
 
-            let claimStore: ClaimsStore = globalPresentableStoreContainer.get()
+            let claimStore: ClaimsStore = hGlobalPresentableStoreContainer.get()
             let claimId = self.claim.id
-            claimStore.stateSignal.plain().publisher
+            claimStore.stateSignal
                 .map({ $0.claim(for: claimId) })
                 .compactMap({ $0?.conversation?.newestMessage?.sentAt })
                 .removeDuplicates()
                 .receive(on: RunLoop.main)
                 .sink { [weak self] value in
-                    let claimStore: ClaimsStore = globalPresentableStoreContainer.get()
+                    let claimStore: ClaimsStore = hGlobalPresentableStoreContainer.get()
                     let conversationId = claimStore.state.claim(for: self?.claim.id ?? "")?.conversation?.id
                     let chatStore: ChatStore = hGlobalPresentableStoreContainer.get()
                     let timeStamp = chatStore.state.conversationsTimeStamp[conversationId ?? ""]
@@ -468,7 +467,7 @@ public class ClaimDetailViewModel: ObservableObject {
 
             Timer.publish(every: 5, on: .main, in: .common).autoconnect()
                 .sink { _ in
-                    let store: ClaimsStore = globalPresentableStoreContainer.get()
+                    let store: ClaimsStore = hGlobalPresentableStoreContainer.get()
                     store.send(.fetchClaims)
                 }
                 .store(in: &cancellables)
