+++ Projects/Claims/Sources/Views/SubmitClaim/Deflect/SubmitClaimDeflectScreen.swift
@@ -1,4 +1,3 @@
-import Presentation
 import SwiftUI
 import hCore
 import hCoreUI
