+++ Projects/MoveFlow/Sources/View/MovingFlowAddressView.swift
@@ -63,7 +63,7 @@ struct MovingFlowAddressView: View {
         }
         .hFormTitle(title: .init(.standard, .displayXSLong, L10n.changeAddressEnterNewAddressTitle))
         .sectionContainerStyle(.transparent)
-        .presentableStoreLensAnimation(.default)
+        .hPresentableStoreLensAnimation(.default)
     }
 
     func addressField() -> some View {
