+++ Projects/Profile/Sources/Views/Components/NotificationsCardView.swift
@@ -30,7 +30,7 @@ struct NotificationsCardView: View {
                     ])
             }
         }
-        .presentableStoreLensAnimation(.default)
+        .hPresentableStoreLensAnimation(.default)
 
     }
 }
