+++ Projects/hCoreUI/Sources/Activity/ActivityIndicatorView.swift
@@ -346,137 +346,6 @@ private struct RetryViewWithError<StoreType: StoreLoading & Store>: ViewModifier
     }
 }
 
-public struct LoadingViewWithGenericError<Content: View, StoreType: StoreLoading & Store>: View {
-    var content: () -> Content
-    @PresentableStore var store: StoreType
-    private let actions: [StoreType.Loading]
-    @Environment(\.presentableStoreLensAnimation) var animation
-    @State var presentError = false
-    @State var error = ""
-    @State var isLoading = false
-    let bottomAction: () -> Void
-    private let showLoading: Bool
-    public init(
-        _ type: StoreType.Type,
-        _ actions: [StoreType.Loading],
-        showLoading: Bool = true,
-        bottomAction: @escaping () -> Void,
-        @ViewBuilder content: @escaping () -> Content
-    ) {
-        self.actions = actions
-        self.content = content
-        self.showLoading = showLoading
-        self.bottomAction = bottomAction
-
-    }
-
-    public var body: some View {
-        ZStack {
-            BackgroundView().edgesIgnoringSafeArea(.all)
-            if isLoading && showLoading {
-                loadingIndicatorView.transition(.opacity.animation(animation ?? .easeInOut(duration: 0.2)))
-            } else if presentError {
-                GenericErrorView(
-                    title: L10n.somethingWentWrong,
-                    description: error,
-                    buttons: .init(
-                        actionButton: .init(buttonAction: {
-                            for action in actions {
-                                store.removeLoading(for: action)
-                            }
-                        }),
-                        dismissButton: .init(
-                            buttonTitle: L10n.openChat,
-                            buttonAction: {
-                                bottomAction()
-                            }
-                        )
-                    )
-                )
-            } else {
-                content().transition(.opacity.animation(animation ?? .easeInOut(duration: 0.2)))
-            }
-        }
-        .onReceive(
-            store.loadingSignal
-                .plain()
-                .publisher
-        ) { value in
-            handle(allActions: value)
-        }
-        .onAppear {
-            let store: StoreType = globalPresentableStoreContainer.get()
-            if let state = handle(allActions: store.loadingSignal.value) {
-                isLoading = state.isLoading
-                error = state.error ?? ""
-                presentError = state.presentError
-            }
-        }
-    }
-
-    @discardableResult
-    private func handle(
-        allActions: [StoreType.Loading: LoadingState<String>]
-    ) -> LoadingViewWithGenericError.ChangeStateValue? {
-        let actions = allActions.filter({ self.actions.contains($0.key) })
-        var state: LoadingViewWithGenericError.ChangeStateValue?
-        if actions.count > 0 {
-            if actions.filter({ $0.value == .loading }).count > 0 {
-                state = .init(isLoading: true, presentError: false, error: nil)
-            } else {
-                var tempError = ""
-                for action in actions {
-                    switch action.value {
-                    case .error(let error):
-                        tempError = error
-                    default:
-                        break
-                    }
-                }
-                state = .init(isLoading: false, presentError: true, error: tempError)
-            }
-        } else {
-            state = .init(isLoading: false, presentError: false, error: nil)
-        }
-        if let state {
-            changeState(to: state)
-        }
-        return state
-    }
-    private func changeState(to state: LoadingViewWithGenericError.ChangeStateValue) {
-        DispatchQueue.main.async {
-            if let animation {
-                withAnimation(animation) {
-                    self.error = state.error ?? ""
-                    self.isLoading = state.isLoading
-                    self.presentError = state.presentError
-                }
-            } else {
-                self.error = state.error ?? ""
-                self.isLoading = state.isLoading
-                self.presentError = state.presentError
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var loadingIndicatorView: some View {
-        HStack {
-            DotsActivityIndicator(.standard)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background(hBackgroundColor.primary.opacity(0.01))
-        .edgesIgnoringSafeArea(.top)
-        .useDarkColor
-    }
-
-    private struct ChangeStateValue {
-        let isLoading: Bool
-        let presentError: Bool
-        let error: String?
-    }
-}
-
 public struct LoadingViewWithContentt: ViewModifier {
     @Binding var isLoading: Bool
     @Binding var error: String?
